Clean up canvas table classes
