+++ studio/LonaStudio/Extensions/NSTableColumn+Init.swift
@@ -1,0 +1,34 @@
+//
+//  NSTableColumn+Init.swift
+//  LonaStudio
+//
+//  Created by Devin Abbott on 12/8/18.
+//  Copyright © 2018 Devin Abbott. All rights reserved.
+//
+
+import AppKit
+
+extension NSTableColumn {
+    convenience init(
+        title: String,
+        resizingMask: ResizingOptions = .autoresizingMask,
+        width: CGFloat? = nil,
+        minWidth: CGFloat? = nil,
+        maxWidth: CGFloat? = nil) {
+        self.init(identifier: NSUserInterfaceItemIdentifier(rawValue: title))
+        self.title = title
+        self.resizingMask = resizingMask
+
+        if let width = width {
+            self.width = width
+        }
+
+        if let minWidth = minWidth {
+            self.minWidth = minWidth
+        }
+
+        if let maxWidth = maxWidth {
+            self.maxWidth = maxWidth
+        }
+    }
+}
